Turn into library + binary with minimal changes
- Rename `try_main` to `run_command` which takes `Vec<CString>` arguments.
  This is the only function exposed by the library, along with its
  possible results (`Exec` and `Error`).

- Make `clap` into an optional (but default) dependency. The library
  can then be reused without `clap`.
