+++ src/main.rs
@@ -1,28 +1,13 @@
-#![deny(unsafe_op_in_unsafe_fn)]
-#![allow(
-    clippy::empty_enum,
-    clippy::let_underscore_untyped,
-    clippy::needless_pass_by_value,
-    clippy::uninlined_format_args
-)]
-
-mod error;
-
-use crate::error::Result;
 use clap::{Arg, ArgAction, Command};
-use nix::pty::{self, ForkptyResult, Winsize};
-use nix::sys::wait::{self, WaitStatus};
-use nix::unistd::{self, Pid};
 use std::ffi::{CString, OsString};
 use std::io::{self, Write};
-use std::os::fd::{AsFd, AsRawFd, BorrowedFd};
 use std::os::unix::ffi::OsStrExt;
 use std::process;
 
-enum Exec {}
+use faketty::run_command;
 
 fn main() -> ! {
-    match try_main() {
+    match run_command(crate::args()) {
         Err(err) => {
             let _ = writeln!(io::stderr(), "faketty: {}", err);
             process::exit(1);
@@ -30,30 +15,6 @@ fn main() -> ! {
         #[allow(unreachable_patterns)]
         Ok(exec) => match exec {},
     }
-}
-
-const STDIN: BorrowedFd = unsafe { BorrowedFd::borrow_raw(0) };
-const STDOUT: BorrowedFd = unsafe { BorrowedFd::borrow_raw(1) };
-const STDERR: BorrowedFd = unsafe { BorrowedFd::borrow_raw(2) };
-
-fn try_main() -> Result<Exec> {
-    let args = crate::args();
-    let new_stdin = STDIN.try_clone_to_owned()?;
-    let new_stderr = STDERR.try_clone_to_owned()?;
-    let pty1 = unsafe { crate::forkpty() }?;
-    if let ForkptyResult::Parent { child, master } = pty1 {
-        crate::copyfd(master.as_fd(), STDOUT);
-        crate::copyexit(child);
-    }
-    let new_stdout = STDOUT.try_clone_to_owned()?;
-    let pty2 = unsafe { crate::forkpty() }?;
-    if let ForkptyResult::Parent { child, master } = pty2 {
-        crate::copyfd(master.as_fd(), new_stderr.as_fd());
-        crate::copyexit(child);
-    }
-    unistd::dup2(new_stdin.as_raw_fd(), STDIN.as_raw_fd())?;
-    unistd::dup2(new_stdout.as_raw_fd(), STDOUT.as_raw_fd())?;
-    crate::exec(args)
 }
 
 fn app() -> Command {
@@ -100,53 +61,6 @@ fn args() -> Vec<CString> {
         .unwrap()
         .map(|os_string| CString::new(os_string.as_bytes()).unwrap())
         .collect()
-}
-
-unsafe fn forkpty() -> Result<ForkptyResult> {
-    let winsize = Winsize {
-        ws_row: 24,
-        ws_col: 80,
-        ws_xpixel: 0,
-        ws_ypixel: 0,
-    };
-    let termios = None;
-    let result = unsafe { pty::forkpty(&winsize, termios) }?;
-    Ok(result)
-}
-
-fn exec(args: Vec<CString>) -> Result<Exec> {
-    let args: Vec<_> = args.iter().map(CString::as_c_str).collect();
-    unistd::execvp(args[0], &args)?;
-    unreachable!();
-}
-
-fn copyfd(read: BorrowedFd, write: BorrowedFd) {
-    const BUF: usize = 4096;
-    let mut buf = [0; BUF];
-    loop {
-        match unistd::read(read.as_raw_fd(), &mut buf) {
-            Ok(0) | Err(_) => return,
-            Ok(n) => {
-                let _ = write_all(write, &buf[..n]);
-            }
-        }
-    }
-}
-
-fn write_all(fd: BorrowedFd, mut buf: &[u8]) -> Result<()> {
-    while !buf.is_empty() {
-        let n = unistd::write(fd, buf)?;
-        buf = &buf[n..];
-    }
-    Ok(())
-}
-
-fn copyexit(child: Pid) -> ! {
-    let flag = None;
-    process::exit(match wait::waitpid(child, flag) {
-        Ok(WaitStatus::Exited(_pid, code)) => code,
-        _ => 0,
-    });
 }
 
 #[test]
